drfc/pyproject.toml
[project]
name = "drfc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "drf-yasg[validation]>=1.21.11",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.2",
]

